+++ packages/components/src/components/propTypes/index.ts
@@ -84,7 +84,6 @@ import type { IllustratedMessageProps } from "@/components/IllustratedMessage";
 import type { RatingProps } from "@/components/Rating";
 import type { CodeEditorProps } from "@/components/CodeEditor";
 import type { KbdProps } from "@/components/Kbd/Kbd";
-import type { TranslationProviderProps } from "@/components/TranslationProvider/TranslationProvider";
 
 export * from "./types";
 
@@ -169,7 +168,6 @@ export interface FlowComponentPropsTypes {
   TextArea: TextAreaProps;
   TextField: TextFieldProps;
   TimeField: TimeFieldProps;
-  TranslationProvider: TranslationProviderProps;
 }
 
 const propsContextSupportingComponentsMap: Record<
@@ -257,7 +255,6 @@ const propsContextSupportingComponentsMap: Record<
   TextArea: true,
   TextField: true,
   TimeField: true,
-  TranslationProvider: true,
 };
 
 export const propsContextSupportingComponents = Object.keys(